Move getters to `parsers.js`
