+++ src/language-css/index.js
@@ -3,6 +3,7 @@
 const createLanguage = require("../utils/create-language");
 const printer = require("./printer-postcss");
 const options = require("./options");
+const parsers = require("./parsers");
 
 const languages = [
   createLanguage(require("linguist-languages/data/CSS.json"), (data) => ({
@@ -36,22 +37,6 @@ const languages = [
 const printers = {
   postcss: printer,
 };
-
-const parsers =
-  process.env.PRETTIER_TARGET === "universal"
-    ? undefined
-    : {
-        // TODO: switch these to just `postcss` and use `language` instead.
-        get css() {
-          return require("./parser-postcss").parsers.css;
-        },
-        get less() {
-          return require("./parser-postcss").parsers.less;
-        },
-        get scss() {
-          return require("./parser-postcss").parsers.scss;
-        },
-      };
 
 module.exports = {
   languages,
