+++ src/language-graphql/parsers.js
@@ -1,0 +1,7 @@
+"use strict";
+
+module.exports = {
+  get graphql() {
+    return require("./parser-graphql").parsers.graphql;
+  },
+};
