+++ src/language-handlebars/index.js
@@ -2,6 +2,7 @@
 
 const createLanguage = require("../utils/create-language");
 const printer = require("./printer-glimmer");
+const parsers = require("./parsers");
 
 const languages = [
   createLanguage(require("linguist-languages/data/Handlebars.json"), () => ({
@@ -14,15 +15,6 @@ const languages = [
 const printers = {
   glimmer: printer,
 };
-
-const parsers =
-  process.env.PRETTIER_TARGET === "universal"
-    ? undefined
-    : {
-        get glimmer() {
-          return require("./parser-glimmer").parsers.glimmer;
-        },
-      };
 
 module.exports = {
   languages,
