+++ src/language-handlebars/parsers.js
@@ -1,0 +1,7 @@
+"use strict";
+
+module.exports = {
+  get glimmer() {
+    return require("./parser-glimmer").parsers.glimmer;
+  },
+};
