+++ src/language-html/index.js
@@ -3,6 +3,7 @@
 const createLanguage = require("../utils/create-language");
 const printer = require("./printer-html");
 const options = require("./options");
+const parsers = require("./parsers");
 
 const languages = [
   createLanguage(require("linguist-languages/data/HTML.json"), () => ({
@@ -40,28 +41,6 @@ const languages = [
 const printers = {
   html: printer,
 };
-
-const parsers =
-  process.env.PRETTIER_TARGET === "universal"
-    ? undefined
-    : {
-        // HTML
-        get html() {
-          return require("./parser-html").parsers.html;
-        },
-        // Vue
-        get vue() {
-          return require("./parser-html").parsers.vue;
-        },
-        // Angular
-        get angular() {
-          return require("./parser-html").parsers.angular;
-        },
-        // Lightning Web Components
-        get lwc() {
-          return require("./parser-html").parsers.lwc;
-        },
-      };
 
 module.exports = {
   languages,
