+++ src/language-yaml/parsers.js
@@ -1,0 +1,7 @@
+"use strict";
+
+module.exports = {
+  get yaml() {
+    return require("./parser-yaml").parsers.yaml;
+  },
+};
